pyproject.toml
[project]
name = "telegram-tutorial-bot"
version = "0.1.0"
description = "Telegram Echo Bot created with the help of Official Telegram Bot Tutorial."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.24.0",
    "python-dotenv>=1.2.1",
]
